+++ packages/editor/web/src/RichContentEditor/RichContentEditor.jsx
@@ -206,12 +206,22 @@ class RichContentEditor extends Component {
   });
 
   getInitialEditorState() {
-    const { editorState, initialState, anchorTarget, relValue } = this.props;
+    const {
+      editorState,
+      initialState,
+      anchorTarget,
+      relValue,
+      normalize: { disableInlineImages = false },
+    } = this.props;
     if (editorState) {
       return editorState;
     }
     if (initialState) {
-      const rawContentState = normalizeInitialState(initialState, { anchorTarget, relValue });
+      const rawContentState = normalizeInitialState(initialState, {
+        anchorTarget,
+        relValue,
+        disableInlineImages,
+      });
       return EditorState.createWithContent(convertFromRaw(rawContentState));
     } else {
       const emptyContentState = convertFromRaw({
@@ -571,6 +581,9 @@ RichContentEditor.propTypes = {
   siteDomain: PropTypes.string,
   iframeDomain: PropTypes.string,
   onError: PropTypes.func,
+  normalize: PropTypes.shape({
+    disableInlineImages: PropTypes.bool,
+  }),
 };
 
 RichContentEditor.defaultProps = {
@@ -581,6 +594,7 @@ RichContentEditor.defaultProps = {
   onError: err => {
     throw err;
   },
+  normalize: {},
 };
 
 export default RichContentEditor;