+++ CHANGELOG.md
@@ -13,6 +13,12 @@
   <summary>
     Changes that have landed in master but are not yet released.
     Click to see more.
+
+### :house: Internal
+- `editor` `viewer`
+  - [#1090](https://github.com/wix-incubator/rich-content/pull/1090) support normalize config with {disableInlineImages: false/true}
+
+
   </summary>